requirements.txt
dnspython==2.6.1
Flask==3.1.1
idna==3.7
Jinja2==3.1.6
python-decouple==3.3
#uWSGI==2.0.25.1 #not needed for local testing
pandas==2.2.2